Split menu logic in many files
